Fix the loaded plugin count.
The addition of support for indicator-ng means that there now are two
ways to load plugins. The main function which loads all functions uses
call-by-reference to get the number of plugins loaded, but the two
functions which do the work overwrite the variable only the plugins
from the second function are counted. Fix this by making the worker
functions add to the variable instead of assigning to it.